1. more warnings fixes in gtk/region.cpp and common/tbarsmpl.cpp 2. wxWindow::Create() always takes a wxValidator parameter, even if    !wxUSE_VALIDATORS - it is just ignored then, tons of #if wxUSE_VALIDATORS    removed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@5275 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/gtk/region.cpp
@@ -349,22 +349,22 @@ void wxRegionIterator::Reset( const wxRegion& region )
 
 wxRegionIterator::operator bool () const
 {
-    return m_current < m_region.GetRectList()->Number();
+    return m_current < (size_t)m_region.GetRectList()->Number();
 }
 
 bool wxRegionIterator::HaveRects() const
 {
-    return m_current < m_region.GetRectList()->Number();
+    return m_current < (size_t)m_region.GetRectList()->Number();
 }
 
 void wxRegionIterator::operator ++ ()
 {
-    if (m_current < m_region.GetRectList()->Number()) ++m_current;
+    if (m_current < (size_t)m_region.GetRectList()->Number()) ++m_current;
 }
 
 void wxRegionIterator::operator ++ (int)
 {
-    if (m_current < m_region.GetRectList()->Number()) ++m_current;
+    if (m_current < (size_t)m_region.GetRectList()->Number()) ++m_current;
 }
 
 wxCoord wxRegionIterator::GetX() const
